Revert "new linter: detect consecutive dashes in filenames and directory paths"
